+++ gen.js
@@ -1,3 +1,4 @@
+import 'dropflow'; // needed because some import below imports WASM
 import fs from 'fs';
 import path from 'path';
 import * as LineBreakTrie from './dist/src/trie-line-break.js';
@@ -473,7 +474,9 @@ const excludedFonts = new Set([
   'Noto Sans HK', // I think TC is more widely applicable
 ]);
 
-const priorities = new Map([ // note 50 is the default priority
+// We only add one of each subset (eg latin, khmer) and there is overlap between
+// families, so tweak the ordering here. Note 50 is the default priority.
+const priorities = new Map([
   ['Noto Sans', 0], // Just because Latin is so common
   ['Noto Sans Arabic', 0], // Noto Sans Kufi Arabic is a traditional style
   ['Noto Sans Math', 0], // Lots of other fonts encode math
@@ -483,58 +486,52 @@ const priorities = new Map([ // note 50 is the default priority
   ['Noto Sans Mongolian', 91], // And Mongolian
 ]);
 
-async function generateSystemFonts() {
+async function generateNotoFonts() {
   const res = await fetch('https://api.fontsource.org/v1/fonts')
-  /** @type {Map<string, string[]>} */
-  const subsetToUrls = new Map();
-  /** @type {Map<string, number>} */
-  const subsetToSubsetId = new Map();
+  /** @type {Set<string>} */
+  const addedSubsets = new Set();
 
   if (res.status !== 200) throw new Error(res.statusText);
 
   const json = await res.json();
-  const trie = new UnicodeTrieBuilder();
+  const calls = [];
+  let skipSubsets = new Set(addedSubsets);
 
-  function addVariants(unicodeRange, variantUrlMaps) {
-    const initialSubsets = new Set(subsetToUrls.keys());
+  function addVariants(unicodeRange, family, weight, style, variantMap) {
+    if (!variantMap) return;
 
-    for (const variantMap of variantUrlMaps) {
-      if (!variantMap) continue;
-      for (const [subset, data] of Object.entries(variantMap)) {
-        const ttf = data?.url?.ttf;
-        if (ttf && !initialSubsets.has(subset) && unicodeRange[subset]) {
-          const subsetUnicodeRange = unicodeRange[subset];
-
-          let urls = subsetToUrls.get(subset);
-          if (!urls) subsetToUrls.set(subset, urls = []);
-
-          let subsetId = subsetToSubsetId.get(subset);
-          if (!subsetId) subsetToSubsetId.set(subset, subsetId = subsetToSubsetId.size + 1);
-
-          for (const range of subsetUnicodeRange.split(',')) {
-            if (range.startsWith('U+')) {
-              const [u1, u2] = range.slice(2).split('-');
-              if (u2) {
-                trie.setRange(parseInt(u1, 16), parseInt(u2, 16), subsetId, false);
-              } else {
-                trie.set(parseInt(u1, 16), subsetId, false);
-              }
-            }
-          }
-
-          console.log(subset, ttf);
-
-          urls.push(ttf);
-        }
+    for (const [subset, data] of Object.entries(variantMap)) {
+      const ttf = data?.url?.ttf;
+      if (ttf && !skipSubsets.has(subset) && unicodeRange[subset]) {
+        const subsetUnicodeRange = unicodeRange[subset];
+        addedSubsets.add(subset);
+        console.log(family, ttf);
+        let ts = '';
+        ts += `fonts.add(\n`;
+        ts += `  new FontFace(\n`;
+        ts += `    '${family}',\n`;
+        ts += `    new URL('${ttf}'),\n`
+        ts += `    {weight: ${weight}, style: '${style}', unicodeRange: '${subsetUnicodeRange}'}\n`
+        ts += `  )\n`;
+        ts += `);\n`;
+        calls.push(ts);
       }
     }
   }
 
+  // Descending priority so it's easy to track which subsets we added already
   json.sort((f1, f2) => {
     const p1 = priorities.get(f1.family) ?? 50;
     const p2 = priorities.get(f2.family) ?? 50;
     return p1 < p2 ? -1 : p1 > p2 ? 1 : 0;
   });
+
+  let ts = '// generated from gen.js\n';
+  ts += 'import {fonts, FontFace} from \'../src/text-font.js\';\n';
+  ts += 'let called = false;\n';
+  ts += 'export default function registerNotoFonts() {\n';
+  ts += 'if (called) return;\n';
+  ts += 'called = true;\n';
 
   for (const {family, category, id} of json) {
     if (
@@ -570,24 +567,23 @@ async function generateSystemFonts() {
       // Some subsetted scripts that aren't Latin have italic versions, but as
       // far as I know, italics is a Latin-specific thing. I looked at
       // Devanagari's italic versions and it was just a bold version.
-      addVariants(unicodeRange, [
-        variants['400']?.['normal'],
-        {latin: variants['400']?.['italic']?.['latin']},
-        variants['700']?.['normal'],
-        {latin: variants['700']?.['italic']?.['latin']}
-      ]);
+      const italicRegular = {latin: variants['400']?.['italic']?.['latin']};
+      const italicBold = {latin: variants['700']?.['italic']?.['latin']};
+
+      skipSubsets = new Set(addedSubsets);
+      addVariants(unicodeRange, family, '400', 'normal', variants['400']?.['normal']);
+      addVariants(unicodeRange, family, '400', 'italic', italicRegular);
+      addVariants(unicodeRange, family, '700', 'normal', variants['700']?.['normal']);
+      addVariants(unicodeRange, family, '700', 'italic', italicBold);
     }
   }
 
-  let ts = '// generated from gen.js\nexport default new Map([\n';
-  for (const [subset, urls] of subsetToUrls) {
-    const subsetId = subsetToSubsetId.get(subset);
-    ts += ` ${JSON.stringify([subsetId, urls])},\n`;
-  }
-  ts += ']);\n';
+  // FontFaceSet is in ascending priority
+  ts += calls.reverse().join('');
 
-  fs.writeFileSync(path.join(__dirname, `gen/system-fonts-database.ts`), ts);
-  writeTrie(path.join(__dirname, 'gen/system-fonts-trie.cc'), 'system_font_trie', trie);
+  ts += '}\n';
+
+  fs.writeFileSync(path.join(__dirname, `gen/register-noto-fonts.ts`), ts);
 }
 
 const fns = process.argv.slice(2).map(command => {
@@ -598,7 +594,7 @@ const fns = process.argv.slice(2).map(command => {
   if (command === 'emoji-trie') return generateEmojiTrie;
   if (command === 'script-trie') return generateScriptTrie;
   if (command === 'script-names') return generateScriptNames;
-  if (command === 'system-fonts') return generateSystemFonts;
+  if (command === 'noto-fonts') return generateNotoFonts;
   if (command === 'derived-core-properties-trie') return generateDerivedCorePropertiesTrie;
   console.error(`Usage: node gen.js (cmd )+
 Available commands:
@@ -609,7 +605,7 @@ Available commands:
   emoji-trie
   script-trie
   script-names
-  system-fonts
+  noto-fonts
   derived-core-properties-trie`);
   process.exit(1);
 });