+++ src/api-register-noto-fonts.ts
@@ -1,0 +1,1 @@
+export {default} from '../gen/register-noto-fonts.js';