+++ src/api.ts
@@ -1,7 +1,7 @@
 import '#register-default-environment';
 import {HTMLElement, TextNode} from './dom.js';
 import {DeclaredStyle, getOriginStyle, computeElementStyle} from './style.js';
-import {fonts, FontFace, getFontUrls, createFaceFromTables} from './text-font.js';
+import {fonts, FontFace, loadFonts, createFaceFromTables} from './text-font.js';
 import {generateBlockContainer, layoutBlockBox, BlockFormattingContext, BlockContainer} from './layout-flow.js';
 import HtmlPaintBackend from './paint-html.js';
 import SvgPaintBackend from './paint-svg.js';
@@ -90,12 +90,13 @@ export function paintToCanvas(root: BlockContainer, ctx: CanvasRenderingContext2
   paint(root, backend);
 }
 
-export function renderToCanvasContext(
+export async function renderToCanvasContext(
   rootElement: HTMLElement,
   ctx: CanvasRenderingContext2D,
   width: number,
   height: number
-) {
+): Promise<void> {
+  await load(rootElement);
   const root = generate(rootElement);
   layout(root, width, height);
   paintToCanvas(root, ctx);
@@ -238,12 +239,7 @@ export function staticLayoutContribution(box: BlockContainer): number {
   return intrinsicSize;
 }
 
-export async function loadNotoFonts(root: HTMLElement): Promise<FontFace[]> {
-  const faces = getFontUrls(root).map(async stringUrl => {
-    const face = await createFaceFromTables(new URL(stringUrl));
-    fonts.add(face);
-    return face;
-  });
-
-  return await Promise.all(faces);
+export async function load(root: HTMLElement): Promise<void> {
+  await loadFonts(root);
+  // TODO: images too
 }